requirements.txt
gspread
google-auth
openai
flask
